+++ Hybrid_GPL_63168/DT-W724V-20140311/package/atp/libraries/opensrc/libupnp/iupnp/inc/service_table.h
@@ -49,7 +49,7 @@ extern "C" {
 #include <stdio.h>
 #include <time.h>
 
-#define SID_SIZE  41
+#define SID_SIZE  (size_t)41
 
 #ifdef INCLUDE_DEVICE_APIS
